sleep 直前のスクリーンショットを撮る & Capybara::ElementNotFound だけでなく StandardError でスクショを通知する
- 合わせて、スクショ通知を NotificationService に切り出す
