+++ README.md
@@ -44,3 +44,24 @@ terraform apply
 # Wake up heroku app before.
 heroku run bundle exec rails start_availability_check
 ```
+
+## 動作確認
+
+ReservationJob の `rescue_from(StandardError)`
+
+```ruby
+require Rails.root.join("domain/models/yokohama/reservation_frame")
+start_date_time = Time.zone.local(2020, 8, 22, 15).to_s
+end_date_time = Time.zone.local(2020, 8, 22, 15).to_s
+
+hash = {
+  park_name: "三ツ沢公園",
+  tennis_court_name: "公園１\nテニスコート１",
+  start_date_time: start_date_time,
+  end_date_time: end_date_time,
+  now: false,
+  state: "will_reserve",
+  id: 1
+}
+ReservationJob.perform_now(hash)
+```
