+++ domain/services/notification_service.rb
@@ -1,8 +1,11 @@
 # frozen_string_literal: true
 
+require "capybara/dsl"
 require Rails.root.join("domain/services/reservation_frames_service")
 
 class NotificationService
+  include Capybara::DSL
+
   def initialize(client = SlackClient.new)
     @client = client
   end
@@ -14,6 +17,12 @@ class NotificationService
 
   def send_message(message)
     client.send(message)
+  end
+
+  def send_screenshot(title, comment)
+    file_path = "tmp/capybara/#{SecureRandom.uuid}.png"
+    save_full_screenshot(file_path)
+    client.upload_png(file_path: file_path, title: title, comment: comment)
   end
 
   private
@@ -35,5 +44,18 @@ class NotificationService
 
   def sort_reservation_frames(reservation_frames)
     ReservationFramesService.new.sort(reservation_frames)
+  end
+
+  def save_full_screenshot(path)
+    # NOTE: https://qiita.com/g-fujioka/items/091c400814800f1280ff
+    # rubocop:disable Layout/LineLength
+    width  = Capybara.page.execute_script("return Math.max(document.body.scrollWidth, document.body.offsetWidth, document.documentElement.clientWidth, document.documentElement.scrollWidth, document.documentElement.offsetWidth);")
+    height = Capybara.page.execute_script("return Math.max(document.body.scrollHeight, document.body.offsetHeight, document.documentElement.clientHeight, document.documentElement.scrollHeight, document.documentElement.offsetHeight);")
+    # rubocop:enable Layout/LineLength
+
+    window = Capybara.current_session.driver.browser.manage.window
+    window.resize_to(width, height)
+
+    page.save_screenshot path
   end
 end
