+++ spec/app/requests/reservation_frames_controller_spec.rb
@@ -38,7 +38,7 @@ RSpec.describe ReservationFramesController, type: :request do
         rf = reservation_frame.to_domain_model
         expect { reserve }.to have_enqueued_job
           .with(rf.to_hash, waiting: true)
-          .at(Date.tomorrow.beginning_of_day + rf.opening_hour.hours - 5.seconds)
+          .at(Date.tomorrow.beginning_of_day + rf.opening_hour.hours - 10.seconds)
       end
 
       it "予約枠の状態を更新する" do
